Feat: Pinned Rooms
This took way too long, but it's finally ready. There is one thing that will need fixing and that is when one room is already open then you try to open it through the pinned rooms list, HackChat does a back-flip and says "no, you can't just do that" so we have to keep track of what rooms are open and reopen the proper page.

Signed-off-by: Liz <lizainslie16@gmail.com>